Fix mapping of saved verbs with the old tense structure to the new tense structure
